refactor: category 관련 코드 위치 롤백
