+++ src/main/java/com/samyookgoo/palgoosam/auction/domain/Category.java
@@ -1,4 +1,4 @@
-package com.samyookgoo.palgoosam.auction.category.domain;
+package com.samyookgoo.palgoosam.auction.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
