+++ src/main/java/com/samyookgoo/palgoosam/auction/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package com.samyookgoo.palgoosam.auction.category.repository;
+package com.samyookgoo.palgoosam.auction.repository;
 
-import com.samyookgoo.palgoosam.auction.category.domain.Category;
+import com.samyookgoo.palgoosam.auction.domain.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
